### 2018-01-09 #498 @wkearn
I could review this submission, but it's not exactly my specialty. I'll try to think of someone more well-versed in these kinds of models and let you know.
