### 2018-01-17 #498 @dvalters
I could review this, although I'm not specifically an ocean modeller (more atmospheric/NWP), but happy to review Fortran/Python/numerical modelling in general.
