### 2018-01-26 #498 @dvalters
Hi @kyleniemeyer - yep I can review
