### 2018-02-20 #592 @whedon
[REVIEW]: Aronnax: An idealised isopycnal ocean model
**Submitting author:** @edoddridge (<a href="http://orcid.org/0000-0002-6097-5729">Edward Doddridge</a>)
**Repository:** <a href="https://github.com/edoddridge/aronnax" target ="_blank">https://github.com/edoddridge/aronnax</a>
**Version:** 0.1.0
