### 2018-02-20 #592 @rabernat
It looks like this issue is specifically for @dvalters' review. Is there a separate one for mine?

First time reviewing for JOSS, sorry for the n00b question...
