### 2018-02-20 #592 @kyleniemeyer
@rabernat no worries—we haven't improved the process yet for two reviewers. We have a single review issue, so that all the interactions take place together. @dvalters will "officially" fill out the checklist, but you should also go through it and see if you agree/disagree with the points; if you do, you can just mention in in a comment.

We also encourage reviewers to file issues in the software's repository (if appropriate), and then mention them here.

For other guidance, our full reviewer guidelines are here: https://joss.theoj.org/about#reviewer_guidelines
