### 2018-03-01 4fe6ef9
Rename `ah` to `kh`
Now the `k` prefix means a thickness diffusivity, as suggested in #176.
