+++ aronnax.f90
@@ -64,7 +64,7 @@ program aronnax
   ! Numerics
   double precision :: dt
   double precision :: au, ar, botDrag
-  double precision :: ah(layerwise_input_length)
+  double precision :: kh(layerwise_input_length)
   double precision :: slip, hmin
   integer          :: niter0, nTimeSteps
   double precision :: dumpFreq, avFreq, checkpointFreq, diagFreq
@@ -123,7 +123,7 @@ program aronnax
   ! TODO Possibly wait until the model is split into multiple files,
   ! then hide the long unsightly code there.
 
-  namelist /NUMERICS/ au, ah, ar, botDrag, dt, slip, &
+  namelist /NUMERICS/ au, kh, ar, botDrag, dt, slip, &
       niter0, nTimeSteps, &
       dumpFreq, avFreq, checkpointFreq, diagFreq, hmin, maxits, & 
       freesurfFac, eps, thickness_error, debug_level
@@ -276,7 +276,7 @@ program aronnax
 
 
   call model_run(h, u, v, eta, depth, dx, dy, wetmask, fu, fv, &
-      dt, au, ar, botDrag, ah, slip, hmin, niter0, nTimeSteps, &
+      dt, au, ar, botDrag, kh, slip, hmin, niter0, nTimeSteps, &
       dumpFreq, avFreq, checkpointFreq, diagFreq, &
       maxits, eps, freesurfFac, thickness_error, &
       debug_level, g_vec, rho0, &
@@ -296,7 +296,7 @@ end program aronnax
 !> Run the model
 
 subroutine model_run(h, u, v, eta, depth, dx, dy, wetmask, fu, fv, &
-    dt, au, ar, botDrag, ah, slip, hmin, niter0, nTimeSteps, &
+    dt, au, ar, botDrag, kh, slip, hmin, niter0, nTimeSteps, &
     dumpFreq, avFreq, checkpointFreq, diagFreq, &
     maxits, eps, freesurfFac, thickness_error, &
     debug_level, g_vec, rho0, &
@@ -327,7 +327,7 @@ subroutine model_run(h, u, v, eta, depth, dx, dy, wetmask, fu, fv, &
   double precision, intent(in) :: fv(0:nx+1, 0:ny+1)
   ! Numerics
   double precision, intent(in) :: dt, au, ar, botDrag
-  double precision, intent(in) :: ah(layers)
+  double precision, intent(in) :: kh(layers)
   double precision, intent(in) :: slip, hmin
   integer,          intent(in) :: niter0, nTimeSteps
   double precision, intent(in) :: dumpFreq, avFreq, checkpointFreq, diagFreq
@@ -534,7 +534,7 @@ subroutine model_run(h, u, v, eta, depth, dx, dy, wetmask, fu, fv, &
     call state_derivative(dhdtveryold, dudtveryold, dvdtveryold, &
         h, u, v, depth, &
         dx, dy, wetmask, hfacW, hfacE, hfacN, hfacS, fu, fv, &
-        au, ar, botDrag, ah, slip, &
+        au, ar, botDrag, kh, slip, &
         RedGrav, g_vec, rho0, wind_x, wind_y, &
         RelativeWind, Cd, &
         spongeHTimeScale, spongeH, &
@@ -550,7 +550,7 @@ subroutine model_run(h, u, v, eta, depth, dx, dy, wetmask, fu, fv, &
     call state_derivative(dhdtveryold, dudtveryold, dvdtveryold, &
         hhalf, uhalf, vhalf, depth, &
[…]
