### 2018-03-02 e61a6e6
Add vertical mass diffusion
A new parameter `kv` allows for diffusive mass fluxes between layers. This is a very naive implementation and cannot cope with arbitrarily thin layers. In most simulations this parameter should be set to 0. Closes #175.

With multiple layers in n layer mode, the asymptotic solution is equal layer thicknesses, rather than the more physically desirable zero stratification.
