### 2018-03-02 ef2fba8
Rename test to have "Hypre" in name
This is used to separate the different compiled versions. Otherwise the code coverage gets confused.
