### 2018-03-02 7e9ae12
Update changelog in docs with vert. diffusion
