+++ docs/development.rst
@@ -5,6 +5,8 @@ Development
 Since latest release
 --------------------
 
+Add vertical diffusion of mass between layers (2 March 2018)
+
 
 Version 0.1.0 (11 December 2017)
 --------------------------------
