### 2018-03-08 59d0a19
Refactor dhdt
In preparation for adding more advection schemes the dhdt subroutine has been refactored to separate:
- horizontal diffusion
- vertical diffsion
- advection
