### 2018-03-08 7f70e33
Add upwind advection shceme and switch to choose
A new parameter `hAdvecScheme` now lets users choose between first order centered and upwind advection schemes.
