+++ aronnax.f90
@@ -72,6 +72,7 @@ program aronnax
   integer maxits
   double precision :: eps, freesurfFac, thickness_error
   integer          :: debug_level
+  integer          :: hAdvecScheme
   ! Model
   double precision :: hmean(layerwise_input_length)
   ! Switch for using n + 1/2 layer physics, or using n layer physics
@@ -124,7 +125,7 @@ program aronnax
   ! then hide the long unsightly code there.
 
   namelist /NUMERICS/ au, kh, kv, ar, botDrag, dt, slip, &
-      niter0, nTimeSteps, &
+      niter0, nTimeSteps, hAdvecScheme, &
       dumpFreq, avFreq, checkpointFreq, diagFreq, hmin, maxits, & 
       freesurfFac, eps, thickness_error, debug_level
 
@@ -153,6 +154,7 @@ program aronnax
   debug_level = 0
   niter0 = 0
   RelativeWind = .FALSE.
+  hAdvecScheme = 1
 
   au = 0d0
   ar = 0d0
@@ -288,7 +290,7 @@ program aronnax
       base_wind_x, base_wind_y, wind_mag_time_series, &
       spongeHTimeScale, spongeUTimeScale, spongeVTimeScale, &
       spongeH, spongeU, spongeV, &
-      nx, ny, layers, RedGrav, DumpWind, &
+      nx, ny, layers, RedGrav, hAdvecScheme, DumpWind, &
       RelativeWind, Cd, &
       MPI_COMM_WORLD, myid, num_procs, ilower, iupper, &
       hypre_grid)
@@ -308,7 +310,7 @@ subroutine model_run(h, u, v, eta, depth, dx, dy, wetmask, fu, fv, &
     base_wind_x, base_wind_y, wind_mag_time_series, &
     spongeHTimeScale, spongeUTimeScale, spongeVTimeScale, &
     spongeH, spongeU, spongeV, &
-    nx, ny, layers, RedGrav, DumpWind, &
+    nx, ny, layers, RedGrav, hAdvecScheme, DumpWind, &
     RelativeWind, Cd, &
     MPI_COMM_WORLD, myid, num_procs, ilower, iupper, &
     hypre_grid)
@@ -357,6 +359,7 @@ subroutine model_run(h, u, v, eta, depth, dx, dy, wetmask, fu, fv, &
   integer,          intent(in) :: nx, ny, layers
   ! Reduced gravity vs n-layer physics
   logical,          intent(in) :: RedGrav
+  integer,          intent(in) :: hAdvecScheme
   ! Whether to write computed wind in the output
   logical,          intent(in) :: DumpWind
   logical,          intent(in) :: RelativeWind
@@ -540,7 +543,7 @@ subroutine model_run(h, u, v, eta, depth, dx, dy, wetmask, fu, fv, &
         h, u, v, depth, &
         dx, dy, wetmask, hfacW, hfacE, hfacN, hfacS, fu, fv, &
         au, ar, botDrag, kh, kv, slip, &
-        RedGrav, g_vec, rho0, wind_x, wind_y, &
+        RedGrav, hAdvecScheme, g_vec, rho0, wind_x, wind_y, &
         RelativeWind, Cd, &
         spongeHTimeScale, spongeH, &
         spongeUTimeScale, spongeU, &
@@ -556,7 +559,7 @@ subroutine model_run(h, u, v, eta, depth, dx, dy, wetmask, fu, fv, &
         hhalf, uhalf, vhalf, depth, &
         dx, dy, wetmask, hfacW, hfacE, hfacN, hfacS, fu, fv, &
[…]
