### 2018-03-08 1d5de30
Test for upwind advection
