### 2018-03-08 2456ea6
Plot difference between blessed and current output
When a test fails, the test suite now produces a plot of the difference between the blessed output and the current output.
