### 2018-03-09 652b283
Refactor code into src directory
This splits the model Fortran file into multiple modules containing themed groupings of subroutines. Also includes a clean command for removing build artefacts.
