+++ src/aronnax.f90
@@ -1,0 +1,228 @@
+!> @author
+!> Ed Doddridge
+!
+!> Aronnax, an idealized isopycnal model with n layers and variable bathymetry.
+!!
+!
+!>
+!>     @mainpage Documentation for aronnax.f90
+!>
+!>     @section Overview
+!>     This model is an isopycnal model on an Arakawa C-grid with n
+!>     layers and arbitrary bathymetry.
+!>
+!>
+!>
+!>    @section Grid
+!>
+!>    /\ ------------
+!>    |  |          |
+!>    |  |          |
+!>    dy U    H     |
+!>    |  |          |
+!>    |  |          |
+!>    \/ Z----V------
+!>        <---dx---->
+!>
+!>    H: tracer point - thickness, Bernoulli potential
+!>    U: velocity point - u and v
+!>    Z: vorticity point - zeta
+!>
+
+
+program aronnax
+
+  use model_main
+  use declarations
+  use mpi
+
+  implicit none
+
+
+  ! TODO Possibly wait until the model is split into multiple files,
+  ! then hide the long unsightly code there.
+
+  namelist /NUMERICS/ au, kh, kv, ar, botDrag, dt, slip, &
+      niter0, nTimeSteps, hAdvecScheme, &
+      dumpFreq, avFreq, checkpointFreq, diagFreq, hmin, maxits, & 
+      freesurfFac, eps, thickness_error, debug_level
+
+  namelist /MODEL/ hmean, depthFile, H0, RedGrav
+
+  namelist /PRESSURE_SOLVER/ nProcX, nProcY
+
+  namelist /SPONGE/ spongeHTimeScaleFile, spongeUTimeScaleFile, &
+      spongeVTimeScaleFile, spongeHfile, spongeUfile, spongeVfile
+
+  namelist /PHYSICS/ g_vec, rho0
+
+  namelist /GRID/ nx, ny, layers, dx, dy, fUfile, fVfile, wetMaskFile
+
+  namelist /INITIAL_CONDITIONS/ initUfile, initVfile, initHfile, initEtaFile
+
+  namelist /EXTERNAL_FORCING/ zonalWindFile, meridionalWindFile, &
+      RelativeWind, Cd, &
+      DumpWind, wind_mag_time_series_file
+
+  ! Set default values here
+  ! io defualt frequencis
+  dumpFreq = 1d9
+  avFreq = 0d0
+  checkpointFreq = 0d0
+  diagFreq = 0d0
+
+
+  debug_level = 0
+  
+  ! start from t = 0
+  niter0 = 0
+
+  ! use N/m^2
+  RelativeWind = .FALSE.
+
+  ! use first-order centered differencing
+  hAdvecScheme = 1
+
+  ! No viscosity or diffusion
+  au = 0d0
+  ar = 0d0
+  kh = 0d0
+  kv = 0d0
+
+
+  open(unit=8, file="parameters.in", status='OLD', recl=80)
+  read(unit=8, nml=NUMERICS)
+  read(unit=8, nml=MODEL)
+  read(unit=8, nml=PRESSURE_SOLVER)
+  read(unit=8, nml=SPONGE)
+  read(unit=8, nml=PHYSICS)
+  read(unit=8, nml=GRID)
+  read(unit=8, nml=INITIAL_CONDITIONS)
+  read(unit=8, nml=EXTERNAL_FORCING)
+  close(unit=8)
+
+
+  ! optionally include the MPI code for parallel runs with external
+  ! pressure solver
+  call MPI_INIT(ierr)
+  call MPI_COMM_RANK(MPI_COMM_WORLD, myid, ierr)
[…]

[…]
