### 2018-03-12 036aa10
Move `Werror` to test opts
And remove standalone compile on Travis.
