### 2018-03-13 4773c17
Create time_stepping module and move RK2 into it
The RK2 algorithm is used to initialise the AB3 timestepping algorithm. This partially implements #182
