+++ src/model_main.f90
@@ -3,6 +3,7 @@ module model_main
   use boundaries
   use barotropic_mode
   use state_deriv
+  use time_stepping
   use enforce_thickness
   implicit none
 
@@ -77,27 +78,21 @@ module model_main
     double precision :: dhdt(0:nx+1, 0:ny+1, layers)
     double precision :: dhdtold(0:nx+1, 0:ny+1, layers)
     double precision :: dhdtveryold(0:nx+1, 0:ny+1, layers)
-    double precision :: hnew(0:nx+1, 0:ny+1, layers)
-    ! for initialisation
-    double precision :: hhalf(0:nx+1, 0:ny+1, layers)
+    double precision :: h_new(0:nx+1, 0:ny+1, layers)
     ! for saving average fields
     double precision :: hav(0:nx+1, 0:ny+1, layers)
 
     double precision :: dudt(0:nx+1, 0:ny+1, layers)
     double precision :: dudtold(0:nx+1, 0:ny+1, layers)
     double precision :: dudtveryold(0:nx+1, 0:ny+1, layers)
-    double precision :: unew(0:nx+1, 0:ny+1, layers)
-    ! for initialisation
-    double precision :: uhalf(0:nx+1, 0:ny+1, layers)
+    double precision :: u_new(0:nx+1, 0:ny+1, layers)
     ! for saving average fields
     double precision :: uav(0:nx+1, 0:ny+1, layers)
 
     double precision :: dvdt(0:nx+1, 0:ny+1, layers)
     double precision :: dvdtold(0:nx+1, 0:ny+1, layers)
     double precision :: dvdtveryold(0:nx+1, 0:ny+1, layers)
-    double precision :: vnew(0:nx+1, 0:ny+1, layers)
-    ! for initialisation
-    double precision :: vhalf(0:nx+1, 0:ny+1, layers)
+    double precision :: v_new(0:nx+1, 0:ny+1, layers)
     ! for saving average fields
     double precision :: vav(0:nx+1, 0:ny+1, layers)
 
@@ -248,9 +243,9 @@ module model_main
       ! Code to work out dhdtveryold, dudtveryold and dvdtveryold
       n = 0
       
-      call state_derivative(dhdtveryold, dudtveryold, dvdtveryold, &
+      call RK2(h_new, u_new, v_new, dhdtveryold, dudtveryold, dvdtveryold, &
           h, u, v, depth, &
-          dx, dy, wetmask, hfacW, hfacE, hfacN, hfacS, fu, fv, &
+          dx, dy, dt, wetmask, hfacW, hfacE, hfacN, hfacS, fu, fv, &
           au, ar, botDrag, kh, kv, slip, &
           RedGrav, hAdvecScheme, g_vec, rho0, wind_x, wind_y, &
           RelativeWind, Cd, &
@@ -259,45 +254,19 @@ module model_main
           spongeVTimeScale, spongeV, &
           nx, ny, layers, n, debug_level)
 
-      ! Calculate the values at half the time interval with Forward Euler
-      hhalf = h+0.5d0*dt*dhdtveryold
-      uhalf = u+0.5d0*dt*dudtveryold
-      vhalf = v+0.5d0*dt*dvdtveryold
+      ! Shuffle arrays: new -> present
+      ! Height and velocity fields
+      h = h_new
+      u = u_new
+      v = v_new
 
-      call state_derivative(dhdtveryold, dudtveryold, dvdtveryold, &
-          hhalf, uhalf, vhalf, depth, &
-          dx, dy, wetmask, hfacW, hfacE, hfacN, hfacS, fu, fv, &
-          au, ar, botDrag, kh, kv, slip, &
-          RedGrav, hAdvecScheme, g_vec, rho0, wind_x, wind_y, &
-          RelativeWind, Cd, &
[…]
