+++ src/model_main.f90
@@ -20,8 +20,8 @@ module model_main
       base_wind_x, base_wind_y, wind_mag_time_series, &
       spongeHTimeScale, spongeUTimeScale, spongeVTimeScale, &
       spongeH, spongeU, spongeV, &
-      nx, ny, layers, RedGrav, hAdvecScheme, DumpWind, &
-      RelativeWind, Cd, &
+      nx, ny, layers, RedGrav, hAdvecScheme, TS_algorithm, AB_order, &
+      DumpWind, RelativeWind, Cd, &
       MPI_COMM_WORLD, myid, num_procs, ilower, iupper, &
       hypre_grid)
     implicit none
@@ -70,28 +70,24 @@ module model_main
     ! Reduced gravity vs n-layer physics
     logical,          intent(in) :: RedGrav
     integer,          intent(in) :: hAdvecScheme
+    integer,          intent(in) :: TS_algorithm
+    integer,          intent(in) :: AB_order
     ! Whether to write computed wind in the output
     logical,          intent(in) :: DumpWind
     logical,          intent(in) :: RelativeWind
     double precision,  intent(in) :: Cd
 
-    double precision :: dhdt(0:nx+1, 0:ny+1, layers)
-    double precision :: dhdtold(0:nx+1, 0:ny+1, layers)
-    double precision :: dhdtveryold(0:nx+1, 0:ny+1, layers)
+    double precision :: dhdt(0:nx+1, 0:ny+1, layers, AB_order)
     double precision :: h_new(0:nx+1, 0:ny+1, layers)
     ! for saving average fields
     double precision :: hav(0:nx+1, 0:ny+1, layers)
 
-    double precision :: dudt(0:nx+1, 0:ny+1, layers)
-    double precision :: dudtold(0:nx+1, 0:ny+1, layers)
-    double precision :: dudtveryold(0:nx+1, 0:ny+1, layers)
+    double precision :: dudt(0:nx+1, 0:ny+1, layers, AB_order)
     double precision :: u_new(0:nx+1, 0:ny+1, layers)
     ! for saving average fields
     double precision :: uav(0:nx+1, 0:ny+1, layers)
 
-    double precision :: dvdt(0:nx+1, 0:ny+1, layers)
-    double precision :: dvdtold(0:nx+1, 0:ny+1, layers)
-    double precision :: dvdtveryold(0:nx+1, 0:ny+1, layers)
+    double precision :: dvdt(0:nx+1, 0:ny+1, layers, AB_order)
     double precision :: v_new(0:nx+1, 0:ny+1, layers)
     ! for saving average fields
     double precision :: vav(0:nx+1, 0:ny+1, layers)
@@ -139,9 +135,6 @@ module model_main
 
     ! Time
     integer*8 :: start_time, last_report_time, cur_time
-
-    ! dummy variable for loading checkpoints
-    character(10)    :: num
 
 
     start_time = time()
@@ -236,102 +229,23 @@ module model_main
     !!!  Initialisation of the model STARTS HERE                            !!!
     !!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!
     if (niter0 .eq. 0) then
-      ! Do two initial time steps with Runge-Kutta second-order.
-      ! These initialisation steps do NOT use or update the free surface.
-      !
-      ! ------------------------- negative 2 time step --------------------------
-      ! Code to work out dhdtveryold, dudtveryold and dvdtveryold
+
       n = 0
-      
-      call RK2(h_new, u_new, v_new, dhdtveryold, dudtveryold, dvdtveryold, &
[…]
