+++ src/time_stepping.f90
@@ -5,6 +5,427 @@ module time_stepping
 
   contains
 
+  
+  ! ---------------------------------------------------------------------------
+  !> Parse the TS_algorithm parameter to determine whether multiple tendency
+  !! arrays need to be stored
+  !! 1 is Forward Euler (part of both AB and RK families)
+  !! 2-9 are AB family algorithms
+  !! 12-19 are RK family algorithms (no need to store multiple tendencies, 
+  !! hence AB_order = 1)
+
+
+  subroutine set_AB_order(TS_algorithm, AB_order)
+    implicit none
+
+    integer,          intent(in)  :: TS_algorithm
+    integer,          intent(out) :: AB_order
+
+    if (TS_algorithm .eq. 1) then
+      ! Forward Euler
+      AB_order = 1
+    else if (TS_algorithm .eq. 2) then
+      ! Second-order AB
+      AB_order = 2
+    else if (TS_algorithm .eq. 3) then
+      ! Third-order AB
+      AB_order = 3
+    else if (TS_algorithm .eq. 4) then
+      ! Fourth-order AB
+      AB_order = 4
+    else if (TS_algorithm .eq. 5) then
+      ! Fifth-order AB
+      AB_order = 5
+    else if (TS_algorithm .eq. 12) then
+      ! Second-order RK
+      AB_order = 1
+    else if (TS_algorithm .eq. 13) then
+      ! Third-order RK
+      AB_order = 1
+    else if (TS_algorithm .eq. 14) then
+      ! Fourth-order RK
+      AB_order = 1
+    else
+      ! TS_algorithm not set correctly
+      call clean_stop(0, .FALSE.)
+    end if
+
+  end subroutine set_AB_order
+
+  
+  ! ---------------------------------------------------------------------------
+  !> Initialise tendencies for linear multi-step timestepping algorithms
+
+  subroutine initialise_tendencies(dhdt, dudt, dvdt, h, u, v, depth, &
+          dx, dy, dt, wetmask, hfacW, hfacE, hfacN, hfacS, fu, fv, &
+          au, ar, botDrag, kh, kv, slip, &
+          RedGrav, hAdvecScheme, AB_order, g_vec, rho0, wind_x, wind_y, &
+          RelativeWind, Cd, &
+          spongeHTimeScale, spongeH, &
+          spongeUTimeScale, spongeU, &
+          spongeVTimeScale, spongeV, &
+          nx, ny, layers, debug_level)
+    implicit none
+
+    double precision, intent(out) :: dhdt(0:nx+1, 0:ny+1, layers, AB_order)
+    double precision, intent(out) :: dudt(0:nx+1, 0:ny+1, layers, AB_order)
+    double precision, intent(out) :: dvdt(0:nx+1, 0:ny+1, layers, AB_order)
+    double precision, intent(inout) :: h(0:nx+1, 0:ny+1, layers)
+    double precision, intent(inout) :: u(0:nx+1, 0:ny+1, layers)
+    double precision, intent(inout) :: v(0:nx+1, 0:ny+1, layers)
+    double precision, intent(in) :: depth(0:nx+1, 0:ny+1, layers)
+    double precision, intent(in) :: dx, dy, dt
+    double precision, intent(in) :: wetmask(0:nx+1, 0:ny+1)
+    double precision, intent(in) :: hfacW(0:nx+1, 0:ny+1)
+    double precision, intent(in) :: hfacE(0:nx+1, 0:ny+1)
[…]
