### 2018-03-14 4665f87
Document new paramter `TS_algorithm`
