### 2018-03-15 680aa42
Split AB family methods into separate module
