+++ src/time_stepping.f90
@@ -1,6 +1,7 @@
 module time_stepping
   use state_deriv
   use boundaries
+  use adams_bashforth
   implicit none
 
   contains
@@ -322,109 +323,6 @@ module time_stepping
     end if
 
   end subroutine timestep
-
-
-! ---------------------------------------------------------------------------
-  !> A first-order Forward Euler algorithm
-  !! This is not a good algorithm. Don't use it, except to show how
-  !! how bad it is.
-
-  subroutine ForwardEuler(X_new, dXdt, X, dt, nx, ny, layers, AB_order)
-    implicit none
-
-    double precision, intent(out) :: X_new(0:nx+1, 0:ny+1, layers)
-    double precision, intent(inout) :: dXdt(0:nx+1, 0:ny+1, layers, AB_order)
-    double precision, intent(in) :: X(0:nx+1, 0:ny+1, layers)
-    double precision, intent(in) :: dt
-    integer,          intent(in) :: nx, ny, layers, AB_order
-
-    X_new = X + dt*dXdt(:,:,:,1)
-
-  end subroutine ForwardEuler
-
-! ---------------------------------------------------------------------------
-  !> A second-order Adams-Bashforth algorithm
-  subroutine AB2(X_new, dXdt, X, dt, nx, ny, layers, AB_order)
-    implicit none
-
-    double precision, intent(out) :: X_new(0:nx+1, 0:ny+1, layers)
-    double precision, intent(inout) :: dXdt(0:nx+1, 0:ny+1, layers, AB_order)
-    double precision, intent(in) :: X(0:nx+1, 0:ny+1, layers)
-    double precision, intent(in) :: dt
-    integer,          intent(in) :: nx, ny, layers, AB_order
-
-    X_new = X + dt*(3d0*dXdt(:,:,:,1) - 1d0*dXdt(:,:,:,2))/2d0
-
-    ! Cycle tendencies
-    dXdt(:,:,:,2) = dXdt(:,:,:,1)
-    
-  end subroutine AB2
-
-! ---------------------------------------------------------------------------
-  !> A third-order Adams-Bashforth algorithm
-  subroutine AB3(X_new, dXdt, X, dt, nx, ny, layers, AB_order)
-    implicit none
-
-    double precision, intent(out) :: X_new(0:nx+1, 0:ny+1, layers)
-    double precision, intent(inout) :: dXdt(0:nx+1, 0:ny+1, layers, AB_order)
-    double precision, intent(in) :: X(0:nx+1, 0:ny+1, layers)
-    double precision, intent(in) :: dt
-    integer,          intent(in) :: nx, ny, layers, AB_order
-
-    X_new = X + dt*(23d0*dXdt(:,:,:,1) - 16d0*dXdt(:,:,:,2) + &
-                    5d0*dXdt(:,:,:,3))/12d0
-
-    ! Cycle tendencies
-    dXdt(:,:,:,3) = dXdt(:,:,:,2)
-    dXdt(:,:,:,2) = dXdt(:,:,:,1)
-    
-  end subroutine AB3
-
-! ---------------------------------------------------------------------------
-  !> A fourth-order Adams-Bashforth algorithm
-  subroutine AB4(X_new, dXdt, X, dt, nx, ny, layers, AB_order)
-    implicit none
-
-    double precision, intent(out) :: X_new(0:nx+1, 0:ny+1, layers)
-    double precision, intent(inout) :: dXdt(0:nx+1, 0:ny+1, layers, AB_order)
[…]
