### 2018-03-15 b7471e9
Document addition of new timestepping algorithms
