+++ docs/development.rst
@@ -5,10 +5,14 @@ Development
 Since latest release
 --------------------
 
+Add Adams-Bashforth family of timestepping algorithms up to fifth-order (15 MArch 2018)
+
 Make test suite plot all differences when a test fails, as suggested in `GH136 <https://github.com/edoddridge/aronnax/issues/136>`_ (12 March 2018)
 
 Refactor Fortran code into modules in src/ directory (12 March 2018)
+
 Add option for first-order upwind advction scheme (8 March 2018)
+
 Add vertical diffusion of mass between layers (2 March 2018)
 
 
