### 2018-03-29 ddfb197
Quick fix to allow Intel compilers to work
