+++ src/adams_bashforth.f90
@@ -12,11 +12,11 @@ module adams_bashforth
   subroutine ForwardEuler(X_new, dXdt, X, dt, nx, ny, layers, AB_order)
     implicit none
 
+    integer,          intent(in) :: nx, ny, layers, AB_order
     double precision, intent(out) :: X_new(0:nx+1, 0:ny+1, layers)
     double precision, intent(inout) :: dXdt(0:nx+1, 0:ny+1, layers, AB_order)
     double precision, intent(in) :: X(0:nx+1, 0:ny+1, layers)
     double precision, intent(in) :: dt
-    integer,          intent(in) :: nx, ny, layers, AB_order
 
     X_new = X + dt*dXdt(:,:,:,1)
 
@@ -27,11 +27,11 @@ module adams_bashforth
   subroutine AB2(X_new, dXdt, X, dt, nx, ny, layers, AB_order)
     implicit none
 
+    integer,          intent(in) :: nx, ny, layers, AB_order
     double precision, intent(out) :: X_new(0:nx+1, 0:ny+1, layers)
     double precision, intent(inout) :: dXdt(0:nx+1, 0:ny+1, layers, AB_order)
     double precision, intent(in) :: X(0:nx+1, 0:ny+1, layers)
     double precision, intent(in) :: dt
-    integer,          intent(in) :: nx, ny, layers, AB_order
 
     X_new = X + dt*(3d0*dXdt(:,:,:,1) - 1d0*dXdt(:,:,:,2))/2d0
 
@@ -45,11 +45,11 @@ module adams_bashforth
   subroutine AB3(X_new, dXdt, X, dt, nx, ny, layers, AB_order)
     implicit none
 
+    integer,          intent(in) :: nx, ny, layers, AB_order
     double precision, intent(out) :: X_new(0:nx+1, 0:ny+1, layers)
     double precision, intent(inout) :: dXdt(0:nx+1, 0:ny+1, layers, AB_order)
     double precision, intent(in) :: X(0:nx+1, 0:ny+1, layers)
     double precision, intent(in) :: dt
-    integer,          intent(in) :: nx, ny, layers, AB_order
 
     X_new = X + dt*(23d0*dXdt(:,:,:,1) - 16d0*dXdt(:,:,:,2) + &
                     5d0*dXdt(:,:,:,3))/12d0
@@ -65,11 +65,11 @@ module adams_bashforth
   subroutine AB4(X_new, dXdt, X, dt, nx, ny, layers, AB_order)
     implicit none
 
+    integer,          intent(in) :: nx, ny, layers, AB_order
     double precision, intent(out) :: X_new(0:nx+1, 0:ny+1, layers)
     double precision, intent(inout) :: dXdt(0:nx+1, 0:ny+1, layers, AB_order)
     double precision, intent(in) :: X(0:nx+1, 0:ny+1, layers)
     double precision, intent(in) :: dt
-    integer,          intent(in) :: nx, ny, layers, AB_order
 
     X_new = X + dt*(55d0*dXdt(:,:,:,1) - 59d0*dXdt(:,:,:,2) + &
                     37d0*dXdt(:,:,:,3) - 9d0*dXdt(:,:,:,4))/24d0
@@ -87,11 +87,11 @@ module adams_bashforth
   subroutine AB5(X_new, dXdt, X, dt, nx, ny, layers, AB_order)
     implicit none
 
+    integer,          intent(in) :: nx, ny, layers, AB_order
     double precision, intent(out) :: X_new(0:nx+1, 0:ny+1, layers)
     double precision, intent(inout) :: dXdt(0:nx+1, 0:ny+1, layers, AB_order)
[…]
