### 2018-04-09 edoddridge/aronnax#186 @edoddridge
Compatibility with Python 3
Aronnax should work with Python 3. Currently it doesn't.

The first error that I find is when running `pytest` - one of the test functions uses a Python 2 style print statement. It's likely that there will be more issues.

If it proves too onerous to support both versions of Python, then we should transition to Python 3.

The test suite should be updated to run the tests under Python 3, either in addition to Python 2, or instead of it.
