### 2018-04-18 #592 @dvalters
Hi @eddodridge, @kyleniemeyer here is my review:

This is going to be quite short as I could find little to fault to prevent it being accepted into JOSS.

The software is engineered to a high standard and the documentation is well written, even as a non ocean modelling specialist I felt I understood what the model did and the basics of how it worked. If the documentation lacks anything, it's a full 'API-style' documentation (i.e. more detailed documentation of the modoules/subroutines in the code). But I don't think this should prevent from being accepted into JOSS now.

I thought the "Reproducing published results" and the "Verification" sections were a particularly strong point of the documentation site. The description of the algorithms behind the model are also much welcomed. The `paper.md` is also nice and detailed.

The use of the pytest testing framework is also well implemented.

I've ticked the 'functionality' box in the checklist above, as it appears to work as expected and produce the expected results. But hopefully the other more domain-specialised reviewer(s) will be able to pick up anything I may have missed in this area.

## Benchmarking/Performance

A question for @edoddridge - is there a test that runs the benchmarking simulations featured in the documentation? (It may be in the test suite and possibly overlooked it...) Just so I can tick off the "Performance" checkbox thoroughly.

## Python 3 compatibility

I was going to comment on the lack of support for Python 3, but I see you already have an issue open here: https://github.com/edoddridge/aronnax/issues/186. It would be good to see this implemented one day due to the impending Python 2.7 retirement soon(ish). Not a barrier to acceptance in JOSS though I don't think.
