### 2018-04-18 #592 @edoddridge
Thanks @dvalters! Really appreciate the feedback.

>  If the documentation lacks anything, it's a full 'API-style' documentation (i.e. more detailed documentation of the modoules/subroutines in the code).

Fair enough. I haven't done this yet because I've been refactoring the code lately, and I'm not sure that it's stable enough to justify the time.

> is there a test that runs the benchmarking simulations featured in the documentation?

There's a script, but it's not part of the test suite, nor is it run as part of the continuous integration (it takes long enough to run that waiting for it every time is a pain); it's `benchmark.py` in the `benchmarks` folder. I'll open a ticket about adding that info to the docs.
