### 2018-04-19 edoddridge/aronnax#187 @edoddridge
Describe benchmark script on 'benchmarking' page in the docs
The [benchmarking section](http://aronnax.readthedocs.io/en/latest/benchmarks.html) of the documentation is based on the output of [benchmarks/benchmark.py](https://github.com/edoddridge/aronnax/blob/b7471e90cce7c2e7195a115739a0e95291bab148/benchmarks/benchmark.py) but the page doesn't say that anywhere. It should.

This was identified in the JOSS review https://github.com/openjournals/joss-reviews/issues/592#issuecomment-382356615
