### 2018-04-20 3076b76
Describe location of benchmarking script
Closes #187
