### 2018-04-21 538e260
Python 3 compatibility fixes #186
Modify:
- iteritems
- loops
- configparser
