### 2018-04-21 57d1c04
Run tests under Python 3.6
