+++ .travis.yml
@@ -2,6 +2,7 @@
 language: python
 python:
   - "2.7"
+  - "3.6"
 
 dist: trusty
 sudo: required
