### 2018-04-25 585f381
Update example, reproduction and benchmark sims
These now work with Python 3, and the latest version of the model.
