### 2018-04-25 ed4067f
Remove underflow flag and update profiling build
Closes #191
