### 2018-04-25 97b27b4
Tidy up declarations and add descriptions
