### 2018-04-26 2fb5e00
Add `wind_depth` variable
This currently does nothing, but will eventually control how deep the wind forcing penetrates.
