+++ src/momentum.f90
@@ -1,5 +1,6 @@
 module momentum
   use boundaries
+  use end_run
   implicit none
 
   contains
@@ -27,55 +28,325 @@ module momentum
     double precision, intent(in)  :: au, ar, slip, dx, dy
     double precision, intent(in)  :: hfacN(0:nx+1, 0:ny+1)
     double precision, intent(in)  :: hfacS(0:nx+1, 0:ny+1)
-    integer, intent(in) :: nx, ny, layers
+    integer,          intent(in)  :: nx, ny, layers
     double precision, intent(in)  :: rho0
     logical,          intent(in)  :: RelativeWind
     double precision, intent(in)  :: Cd
     double precision, intent(in)  :: spongeTimeScale(0:nx+1, 0:ny+1, layers)
     double precision, intent(in)  :: spongeU(0:nx+1, 0:ny+1, layers)
-    logical, intent(in) :: RedGrav
+    logical,          intent(in)  :: RedGrav
     double precision, intent(in)  :: botDrag
+
+    integer          :: i, j, k
+    double precision :: dudt_visc(0:nx+1, 0:ny+1, layers)
+    double precision :: dudt_vort(0:nx+1, 0:ny+1, layers)
+    double precision :: dudt_BP(0:nx+1, 0:ny+1, layers)
+    double precision :: dudt_sponge(0:nx+1, 0:ny+1, layers)
+    double precision :: dudt_wind(0:nx+1, 0:ny+1, layers)
+    double precision :: dudt_drag(0:nx+1, 0:ny+1, layers)
+
+
+    dudt = 0d0
+    dudt_visc = 0d0
+    dudt_vort = 0d0
+    dudt_BP = 0d0
+    dudt_sponge = 0d0
+    dudt_wind = 0d0
+    dudt_drag = 0d0
+
+    call evaluate_dudt_visc(dudt_visc, h, u, au, slip, dx, dy, hfacN, &
+      hfacS, nx, ny, layers)
+
+    call evaluate_dudt_vort(dudt_vort, v, zeta, fu, nx, ny, layers)
+
+    call evaluate_dudt_BP(dudt_BP, b, dx, nx, ny, layers)
+
+    call evaluate_dudt_sponge(dudt_sponge, u, spongeTimeScale, spongeU, &
+      nx, ny, layers)
+
+    call evaluate_dudt_wind(dudt_wind, h, u, v, wind_x, wind_y, &
+      wind_depth, nx, ny, layers, rho0, RelativeWind, Cd)
+
+    call evaluate_dudt_drag(dudt_drag, u, ar, nx, ny, layers, RedGrav, &
+      botDrag)
+
+    dudt = dudt + dudt_visc + dudt_vort + dudt_BP + dudt_sponge + &
+            dudt_wind + dudt_drag
+
+    call wrap_fields_3D(dudt, nx, ny, layers)
+
+    return
+  end subroutine evaluate_dudt
+
+
+  ! ---------------------------------------------------------------------------
+  ! Calculate the viscous tendency of zonal velocity for each of the
+  !   active layers
+
+  subroutine evaluate_dudt_visc(dudt_visc, h, u, au, slip, dx, dy, hfacN, &
+      hfacS, nx, ny, layers)
+    implicit none
+
+    ! dudt_visc(i, j) is evaluated at the centre of the left edge of the grid
+    ! box, the same place as u(i, j).
+    double precision, intent(out) :: dudt_visc(0:nx+1, 0:ny+1, layers)
+    double precision, intent(in)  :: h(0:nx+1, 0:ny+1, layers)
+    double precision, intent(in)  :: u(0:nx+1, 0:ny+1, layers)
[…]
