+++ src/time_stepping.f90
@@ -59,7 +59,7 @@ module time_stepping
 
   subroutine initialise_tendencies(dhdt, dudt, dvdt, h, u, v, depth, &
           dx, dy, dt, wetmask, hfacW, hfacE, hfacN, hfacS, fu, fv, &
-          au, ar, botDrag, kh, kv, slip, &
+          au, ar, botDrag, kh, kv, hmin, slip, &
           RedGrav, hAdvecScheme, AB_order, g_vec, rho0, wind_x, wind_y, &
           wind_depth, RelativeWind, Cd, &
           spongeHTimeScale, spongeH, &
@@ -85,6 +85,7 @@ module time_stepping
     double precision, intent(in) :: fv(0:nx+1, 0:ny+1)
     double precision, intent(in) :: au, ar, botDrag
     double precision, intent(in) :: kh(layers), kv
+    double precision, intent(in) :: hmin
     double precision, intent(in) :: slip
     logical,          intent(in) :: RedGrav
     integer,          intent(in) :: hAdvecScheme
@@ -121,7 +122,7 @@ module time_stepping
           dhdt(:,:,:,AB_order-t), dudt(:,:,:,AB_order-t), dvdt(:,:,:,AB_order-t), &
           h, u, v, depth, &
           dx, dy, dt, wetmask, hfacW, hfacE, hfacN, hfacS, fu, fv, &
-          au, ar, botDrag, kh, kv, slip, &
+          au, ar, botDrag, kh, kv, hmin, slip, &
           RedGrav, hAdvecScheme, g_vec, rho0, wind_x, wind_y, &
           wind_depth, RelativeWind, Cd, &
           spongeHTimeScale, spongeH, &
@@ -157,7 +158,7 @@ module time_stepping
   subroutine timestep(h_new, u_new, v_new, dhdt, dudt, dvdt, &
       h, u, v, depth, &
       dx, dy, dt, wetmask, hfacW, hfacE, hfacN, hfacS, fu, fv, &
-      au, ar, botDrag, kh, kv, slip, &
+      au, ar, botDrag, kh, kv, hmin, slip, &
       RedGrav, hAdvecScheme, TS_algorithm, AB_order, &
       g_vec, rho0, wind_x, wind_y, &
       wind_depth, RelativeWind, Cd, &
@@ -188,6 +189,7 @@ module time_stepping
     double precision, intent(in)    :: fv(0:nx+1, 0:ny+1)
     double precision, intent(in)    :: au, ar, botDrag
     double precision, intent(in)    :: kh(layers), kv
+    double precision, intent(in)    :: hmin
     double precision, intent(in)    :: slip
     logical,          intent(in)    :: RedGrav
     integer,          intent(in)    :: hAdvecScheme, TS_algorithm, AB_order
@@ -214,7 +216,7 @@ module time_stepping
       ! Forward Euler
       call state_derivative(dhdt(:,:,:,1), dudt(:,:,:,1), dvdt(:,:,:,1), h, u, v, depth, &
           dx, dy, wetmask, hfacW, hfacE, hfacN, hfacS, fu, fv, &
-          au, ar, botDrag, kh, kv, slip, &
+          au, ar, botDrag, kh, kv, hmin, slip, &
           RedGrav, hAdvecScheme, g_vec, rho0, wind_x, wind_y, &
           wind_depth, RelativeWind, Cd, &
           spongeHTimeScale, spongeH, &
@@ -232,7 +234,7 @@ module time_stepping
       ! Second-order AB
       call state_derivative(dhdt(:,:,:,1), dudt(:,:,:,1), dvdt(:,:,:,1), h, u, v, depth, &
[…]
