### 2018-04-27 bf9ab3c
Add n-layer outcropping test
Runs a rentrant channel with two layers and varaible bathymetry. Each layer starts with a thickness of 0.1 m at some latitude in the channel.
