### 2018-04-30 633be55
Add discussion of outcropping
