### 2018-04-30 8f346fa
Document code changes that allow outcropping
