+++ docs/development.rst
@@ -5,6 +5,7 @@ Development
 Since latest release
 --------------------
 
+- Allow outcropping `GH196 <https://github.com/edoddridge/aronnax/pull/196>`_ (30 April 2018)
 - Add Python 3 compatibility `GH189 <https://github.com/edoddridge/aronnax/pull/189>`_ (23 April 2018)
 - Add Adams-Bashforth family of timestepping algorithms up to fifth-order `GH184 <https://github.com/edoddridge/aronnax/pull/184>`_ (15 March 2018)
 - Make test suite plot all differences when a test fails, as suggested in `GH136 <https://github.com/edoddridge/aronnax/issues/136>`_ and implemented in `GH181 <https://github.com/edoddridge/aronnax/pull/181>`_ (12 March 2018)
