### 2018-05-01 #592 @dvalters
Thanks @edoddridge - I am happy with it all from my point of view as reviewer
