### 2018-05-02 64c07dc
Add Yang et al. (2016) reproduction
