+++ reproductions/Yang_et_al_2016/plot_spin_up.py
@@ -1,0 +1,176 @@
+import os
+import os.path as p
+
+import glob
+from builtins import range
+
+import numpy as np
+import matplotlib
+matplotlib.use("Agg")
+import matplotlib.pyplot as plt
+
+self_path = p.dirname(p.abspath(__file__))
+root_path = p.dirname(self_path)
+
+import sys
+sys.path.append(p.join(root_path, 'test'))
+sys.path.append(p.join(root_path, 'reproductions/Yang_et_al_2016'))
+
+import subprocess as sub
+
+import aronnax as aro
+
+
+## Pretty plots
+plt.rcParams['figure.figsize'] = (12, 12) # set default figure size to 12x12 inches
+plt.rc('text',usetex=True)
+#font = {'family':'serif','size':16}
+font = {'family':'serif','size':16, 'serif': ['computer modern roman']}
+plt.rc('font',**font)
+plt.rc('legend',**{'fontsize':14})
+matplotlib.rcParams['text.latex.preamble']=[r'\usepackage{amsmath}']
+
+
+def wetmask(X, Y):
+    """The wet mask."""
+
+    # start with land everywhere and carve out space for water
+    wetmask = np.zeros(X.shape, dtype=np.float64)
+    # circular gyre region
+    wetmask[((Y-500)**2 + (X-500)**2) < 500**2] = 1
+
+    # clean up the edges
+    wetmask[ 0, :] = 0
+    wetmask[-1, :] = 0
+    wetmask[ :, 0] = 0
+    wetmask[ :,-1] = 0
+
+    return wetmask
+
+nx = 200
+ny = 200
+layers = 2
+xlen = 1000e3
+ylen = 1000e3
+dx = xlen / nx
+dy = ylen / ny
+
+grid = aro.Grid(nx, ny, layers, dx, dy)
+
+X_h, Y_h = np.meshgrid(grid.x/1e3, grid.y/1e3)
+mask_h = wetmask(X_h, Y_h)
+X_v, Y_v = np.meshgrid(grid.x/1e3, grid.yp1/1e3)
+mask_v = wetmask(X_v, Y_v)
+
+def plt_h_cross_section(simulation=None):
+    h_files = sorted(glob.glob("{0}output/snap.h.*".format(simulation)))
+    h_final = aro.interpret_raw_file(h_files[-1], nx, ny, layers)
+
+    plt.figure()
+    for i in range(100,160):
+        plt.plot(-h_final[:,i,0])
+    plt.title('Timestep {}: Day {:05d}'.format(h_files[-1][-10:], len(h_files)))
+    plt.savefig('{0}figures/h_final.png'.format(simulation))
+    plt.close()
+
+    plt.pcolormesh(X_h, Y_h, np.transpose(h_final[:,:,0]))
+    plt.colorbar()
+    plt.contour(X_h, Y_h, np.transpose(h_final[:,:,0]), np.arange(0,800,50), colors='k')
+    plt.title('Timestep {}: Day {:05d}'.format(h_files[-1][-10:], len(h_files)))
+    plt.savefig('{0}figures/h_final_pcolor.png'.format(simulation))
+    plt.close()
+
+
+def plt_state(simulation=None):
+    h_files = sorted(glob.glob("{0}output/snap.h.*".format(simulation)))
+    v_files = sorted(glob.glob("{0}output/snap.v.*".format(simulation)))
+
+    h_max = np.zeros(len(h_files))
+
+    for i in range(len(v_files)):
+        h = aro.interpret_raw_file(h_files[i], nx, ny, layers)
+        h_max[i] = h[100,100,0] #np.max(h[:,:,0])
+    
+    years = np.arange(len(h_max))/360
+    plt.figure()
+    plt.plot(years, h_max)
[…]
