### 2018-05-07 8897e84
Remove transpose from interpret_raw_files
This leaves the axis ordering as (z,y,x), which is consistent with Commodo conventions, and closes #197.
