+++ test/physics_tests.py
@@ -109,10 +109,10 @@ def f_plane_init_u_test(physics, aro_exec, dt):
             # plt.savefig('h.{0}.png'.format(ufile[-10:]),dpi=150)
             # plt.close()
 
-            energy[counter] = nx * ny * (dx * dy * rho0 * (np.sum(np.absolute(h * (u[1:,...]**2 + u[:-1,...]**2)/4.)/2.) + np.sum(np.absolute(h * (v[:,1:,:]**2 + v[:,:-1,:]**2)/4.)/2.)) +  
+            energy[counter] = nx * ny * (dx * dy * rho0 * (np.sum(np.absolute(h * (u[:,:,1:]**2 + u[:,:,:-1]**2)/4.)/2.) + np.sum(np.absolute(h * (v[:,1:,:]**2 + v[:,:-1,:]**2)/4.)/2.)) +  
               dx * dy * rho0 * 0.01 * np.sum(np.absolute(h - 400.)))
 
-            momentum[counter] = nx * ny * dx * dy * rho0 * (np.sum(np.absolute(h * (u[1:,...] + u[:-1,...])/2.)) + np.sum(np.absolute(h * (v[:,1:,:] + v[:,:-1,:])/2.)))
+            momentum[counter] = nx * ny * dx * dy * rho0 * (np.sum(np.absolute(h * (u[:,:,1:] + u[:,:,:-1])/2.)) + np.sum(np.absolute(h * (v[:,1:,:] + v[:,:-1,:])/2.)))
 
             volume[counter] = np.sum(h)
 
@@ -130,9 +130,9 @@ def f_plane_init_u_test(physics, aro_exec, dt):
         X, Y = np.meshgrid(grid.x, grid.yp1)
         init_v = aro.interpret_raw_file(vfiles[1], nx, ny, layers) #init_V(X, Y, True)
 
-        energy_expected[:] = nx * ny * (dx * dy * rho0 * (np.sum(np.absolute(400. * (init_u[1:,...]**2 + init_u[:-1,...]**2)/4.)/2.) + np.sum(np.absolute(400. * (init_v[:,1:,:]**2 + init_v[:,:-1,:]**2)/4.)/2.)) )
+        energy_expected[:] = nx * ny * (dx * dy * rho0 * (np.sum(np.absolute(400. * (init_u[:,:,1:]**2 + init_u[:,:,:-1]**2)/4.)/2.) + np.sum(np.absolute(400. * (init_v[:,1:,:]**2 + init_v[:,:-1,:]**2)/4.)/2.)) )
 
-        momentum_expected[:] = nx * ny * dx * dy * rho0 * (np.sum(np.absolute(h * (init_u[1:,...] + init_u[:-1,...])/2.)) + np.sum(np.absolute(h * (init_v[:,1:,:] + init_v[:,:-1,:])/2.)))
+        momentum_expected[:] = nx * ny * dx * dy * rho0 * (np.sum(np.absolute(h * (init_u[:,:,1:] + init_u[:,:,:-1])/2.)) + np.sum(np.absolute(h * (init_v[:,1:,:] + init_v[:,:-1,:])/2.)))
 
         # print momentum[0]/momentum_expected[0]
 
@@ -271,7 +271,7 @@ def f_plane_wind_test(physics, aro_exec, nx, ny, dx, dy, dt, nTimeSteps):
             # plt.savefig('h.{0}.png'.format(ufile[-10:]),dpi=150)
             # plt.close()
 
[…]
