### 2018-05-07 faf068a
Python 2-> 3 compatability
