### 2018-05-30 #592 @rabernat
I will look it over later today.

Sent from my iPhone

> On May 30, 2018, at 10:57 AM, Ed Doddridge <notifications@github.com> wrote:
>
> @kyleniemeyer @rabernat - any news on this?
>
