### 2018-06-13 #592 @rabernat
Sorry for my long delay. I am satisfied by @dvalters's detailed review and am happy to approve this publication.
