### 2018-06-15 90a838b
Implement funciton to lazily load multiple outputs
