### 2018-06-15 #592 @edoddridge
Thanks @rabernat and @kyleniemeyer.

I've updated the author affiliations, and the DOI is 10.5281/zenodo.1290742
