### 2018-06-15 #592 @kyleniemeyer
@arfon this paper is now accepted and ready to be published!
