### 2018-06-15 #592 @arfon
@dvalters & @rabernat - many thanks for your review here and to @kyleniemeyer for editing this submission ✨

@edoddridge - your paper is now published in JOSS and your DOI is https://doi.org/10.21105/joss.00592 :zap: :rocket: :boom:
